Versiones (informes)/ppi_10 - Informe 3/ppi_10/requirements.txt
asgiref==3.6.0
Django==4.1.7
sqlparse==0.4.3
tzdata==2022.7
django-taggit==3.1.0